+++ R/datasets.R
@@ -1,11 +1,12 @@
 #' outcomerate Formula Matrix (Internal Data)
 #'
 #' The `fmat` object is the internal dataset used by the `outcomerate` package.
-#' It holds all definitions for the outcome rates. With the exception of location
-#' rates, these are taken from the AAPOR Standard Definitions (2016).
+#' It holds all definitions for the outcome rates. With the exception of
+#' location rates, these are taken from the AAPOR Standard Definitions (2023).
 #'
 #' The data is a 3-dimensional binary array consisting of:
-#'   * outcome: codes {I, P, R, NC, O, UH, UO, eUH, eUO, NE}
+#'   * outcome: codes `I`, `P`, `R`, `NC`, `O`, `UH`, `UR`, `UO`, `eUH`,
+#'   `eUR`, `eUO`, and `NE`
 #'   * rate: the shorthand name for the rate (e.g. RR1)
 #'   * side: numerator (NUM) and denominator (DEN)
 #'
@@ -13,15 +14,16 @@
 #' number (i.e. a fraction) consisting of a summation of frequencies of
 #' outcome codes (where the matrix entries are nonzero).
 #'
-#' The input parameters given by the user are {I, P, R, NC, O, UH, UO} and
-#' the parameter 'e'. The parameter e is multiplied by {UH, UO} internally so as to
-#' produce {eUH, eUO}.
+#' The input parameters given by the user are `I`, `P`, `R`, `NC`, `O`, `UH`,
+#' `UR`, `UO`, and `NE`. The scalar or category-specific values in `e` are
+#' multiplied by `UH`, `UR`, and `UO` internally to produce the estimated
+#' eligible counts `eUH`, `eUR`, and `eUO`.
 #'
 #' The reason for this implementation is:
 #'
 #' a) It conforms to a DRY (don't repeat yourself) philosophy by
-#'  holding all definitions in one place. These definitions can be used as upstream
-#'  inputs to functions/test suites requiring them.
+#'  holding all definitions in one place. These definitions can be used as
+#'  upstream inputs to functions/test suites requiring them.
 #'
 #' b) It makes it easier to use intermediate steps in the formula calculations.
 #'  For instance, it may be of use to a researchers to want to obtain the
@@ -34,7 +36,8 @@
 #'
 #' @name fmat
 #' @docType data
-#' @references \url{https://www.aapor.org/Standards-Ethics/Standard-Definitions-(1).aspx}
+#' @references
+#'   \url{https://aapor.org/wp-content/uploads/2024/03/Standards-Definitions-10th-edition.pdf}
 #' @keywords data
 #' @examples
 #' fmat <- outcomerate:::fmat
@@ -42,10 +45,10 @@
 #' # Print the dimensions
 #' dimnames(fmat)
 #'
-#' # Say we want to know the defintion of Response Rate 2, RR2. We see
+#' # Say we want to know the definition of Response Rate 2, RR2. We see
 #' # below that the numerator (NUM) column is defined by the entries with a 1,
 #' # or (I + P). Likewise, the denominator (DEN) is defined as
-#' # (I + P + R + NC + O + UH + UO)
+#' # (I + P + R + NC + O + UH + UR + UO)
 #' fmat[, "RR2", ]
 #'
 #'
@@ -55,7 +58,8 @@
 #' N = fmat[ , , 1]
 #' D = fmat[ , , 2]
 #' x <- c(I = 5, P = 2, R = 1, NC = 7, O = 3,
-#'       UH = 4, UO = 8,  NE = 1, eUH = 3, eUO = 6)
+#'       UH = 4, UR = 2, UO = 8, NE = 1,
+#'       eUH = 3, eUR = 1.5, eUO = 6)
 #'
 #' # Return all rates
 #' (x %*% N) / (x %*% D)
@@ -73,14 +77,17 @@ NULL
 #' conducted in J.R.R. Tolkien's fictional world of Middle Earth.
 #'
 #' Variables contained in the data:
-#'   * __code:__ one of the outcome codes {I, P, R, NC, O, UH, UO, UH, UO, NE}
+#'   * __code:__ one of the outcome codes `I`, `P`, `R`, `NC`, `O`, `UH`,
+#'   `UO`, and `NE`
 #'   * __outcome:__ A human-interpretable label for the `code` variable
 #'   * __researcher__: An identifier for the researcher conducting the interview
 #'   * __region__: The region of the respondent (one of five)
 #'   * __Q1__: A hypothetical binary research question posed to respondents
 #'   * __Q2__: A hypothetical continuous scale question posed to respondents
-#'   * __day__: The day the interview took place (1 being the first day of fieldwork)
-#'   * __race__: The race of the respondent in middle earth (Dwarf, Elf, Hobbit, Man, or Wizard)
+#'   * __day__: The day the interview took place (1 being the first day of
+#'   fieldwork)
+#'   * __race__: The race of the respondent in middle earth (Dwarf, Elf, Hobbit,
+#'   Man, or Wizard)
 #'   * __svywt__: The survey weight (inverse probability of selection)
 #'
 #' @name middleearth